requirements.txt
google-generativeai
google-genai
tavily-python
yfinance
python-dotenv
fastapi
uvicorn
requests
openai
